+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import BooksList from './BooksList';
+import BooksList from '../containers/BooksList';
 import BooksForm from '../containers/BooksForm';
 
 const App = () => (
